+++ domain/src/main/java/joist/domain/orm/impl/InstanceInserter.java
@@ -69,12 +69,12 @@ public class InstanceInserter<T extends DomainObject> {
       for (T instance : instances) {
         step.addParametersForInstance(allParameters, instance);
       }
-      Integer[] keys = Jdbc.insertBatch(UoW.getConnection(), step.sql, allParameters);
+      Long[] keys = Jdbc.insertBatch(UoW.getConnection(), step.sql, allParameters);
       if (step == this.stepsNewId.get(0)) {
         for (int i = 0; i < keys.length; i++) {
           T instance = instances.get(i);
           AliasRegistry.get(instance).getIdColumn().setJdbcValue(instance, keys[i]);
-          AliasRegistry.get(instance).getVersionColumn().setJdbcValue(instance, 0);
+          AliasRegistry.get(instance).getVersionColumn().setJdbcValue(instance, 0l);
           UoW.getIdentityMap().store(instance);
         }
       }
